+++ src/components/modal/Modal.test.jsx
@@ -11,43 +11,43 @@ describe('Modal test', () => {
 
     it('create', () => {
         const modal = mount(
-          <Modal
-            title="Tips"
-            size="tiny"
-            visible={false}
-            lockScroll={false}
+            <Modal
+                title="Tips"
+                size="tiny"
+                visible={false}
+                lockScroll={false}
             >
-            <ModalBody>
-              <span>This is a message</span>
-            </ModalBody>
-            <ModalFooter className="modal-footer">
-              <Button>Cancel</Button>
-              <Button type="primary">Confirm</Button>
-            </ModalFooter>
-          </Modal>
+                <ModalBody>
+                    <span>This is a message</span>
+                </ModalBody>
+                <ModalFooter className="modal-footer">
+                    <Button>Cancel</Button>
+                    <Button type="primary">Confirm</Button>
+                </ModalFooter>
+            </Modal>
         );
 
         expect(modal.find('.cd-modal').exists()).toBe(true);
         modal.find('.cd-modal__wrapper').simulate('click');
-        modal.find(".cd-modal__wrapper").simulate('keydown', {
+        modal.find('.cd-modal__wrapper').simulate('keydown', {
             keyCode: 27
         });
 
         modal.setProps({
-            visible: true
+                           visible: true
                        });
 
         modal.setProps({
-            lockScroll: true,
-            visible: false
+                           lockScroll: true,
+                           visible: false
                        });
 
         modal.setProps({
-            visible: true
+                           visible: true
                        });
 
         modal.setProps({
-            visible: false
+                           visible: false
                        });
 
         modal.unmount();
@@ -56,21 +56,21 @@ describe('Modal test', () => {
     it('create', () => {
         const onClose = sinon.spy();
         const modal = mount(
-          <Modal
-            title="Tips"
-            size="tiny"
-            visible={false}
-            lockScroll={false}
-            onCancel={onClose}
+            <Modal
+                title="Tips"
+                size="tiny"
+                visible={false}
+                lockScroll={false}
+                onCancel={onClose}
             >
-            <ModalBody>
-              <span>This is a message</span>
-            </ModalBody>
-            <ModalFooter className="modal-footer">
-              <Button>Cancel</Button>
-              <Button type="primary">Confirm</Button>
-            </ModalFooter>
-          </Modal>
+                <ModalBody>
+                    <span>This is a message</span>
+                </ModalBody>
+                <ModalFooter className="modal-footer">
+                    <Button>Cancel</Button>
+                    <Button type="primary">Confirm</Button>
+                </ModalFooter>
+            </Modal>
         );
 
         expect(modal.find('.cd-modal').exists()).toBe(true);