+++ src/components/modal/ModalBody.jsx
@@ -4,11 +4,11 @@ import React from 'react';
 import { Component } from '../../../libs';
 
 export default class ModalBody extends Component {
-  render(): React.DOM {
-    return (
-      <div style={this.style()} className={this.className('cd-modal__body')}>
-        { this.props.children }
-      </div>
-    )
-  }
+    render(): React.DOM {
+        return (
+            <div style={this.style()} className={this.className('cd-modal__body')}>
+                {this.props.children}
+            </div>
+        )
+    }
 }