+++ src/components/modal/docs/index.jsx
@@ -1,7 +1,7 @@
 import Markdown from '../../../../libs/markdown';
 
 export default class Modal extends Markdown {
-  document() {
-    return require(`./Modal.md`);
-  }
+    document() {
+        return require(`./Modal.md`);
+    }
 }