+++ src/components/select/docs/index.jsx
@@ -1,0 +1,7 @@
+import Markdown from '../../../../libs/markdown';
+
+export default class Select extends Markdown {
+  document() {
+    return require(`./Select.md`);
+  }
+}