+++ src/index.js
@@ -12,6 +12,7 @@ export { default as Main } from './components/main';
 export { default as Menu } from './components/menu';
 export { default as StepBar } from './components/step-bar';
 export { default as Tag } from './components/tag';
+export { default as Select } from './components/select';
 export { default as Popover } from './components/popover';
 export { default as Radio } from './components/radio';
 export { Transition } from '../libs';