+++ CHANGELOG.md
@@ -25,3 +25,4 @@ Formato de CHANGELOG baseado em [Keep a Changelog](https://keepachangelog.com/pt
 - Componente `Modal`
 - Componente `Popover`
 - Componente `Radio`
+- Componente `Select`